NOJIRA forenkle kodeverkklient
