--- integrasjon/kodeverk-klient/src/main/java/no/nav/vedtak/felles/integrasjon/kodeverk/KodeverkConsumerProducerDelegator.java
@@ -1,25 +1,0 @@
-package no.nav.vedtak.felles.integrasjon.kodeverk;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.enterprise.inject.Produces;
-import javax.inject.Inject;
-
-@ApplicationScoped
-public class KodeverkConsumerProducerDelegator {
-    private KodeverkConsumerProducer producer;
-
-    @Inject
-    public KodeverkConsumerProducerDelegator(KodeverkConsumerProducer producer) {
-        this.producer = producer;
-    }
-
-    @Produces
-    public KodeverkConsumer kodeverkConsumerForEndUser() {
-        return producer.kodeverkConsumer();
-    }
-
-    @Produces
-    public KodeverkSelftestConsumer kodeverkSelftestConsumerForSystemUser() {
-        return producer.kodeverkSelftestConsumer();
-    }
-}
