--- integrasjon/kodeverk-klient/src/main/java/no/nav/vedtak/felles/integrasjon/kodeverk/KodeverkSelftestConsumer.java
@@ -1,6 +1,0 @@
-package no.nav.vedtak.felles.integrasjon.kodeverk;
-
-import no.nav.vedtak.felles.integrasjon.felles.ws.SelftestConsumer;
-
-public interface KodeverkSelftestConsumer extends SelftestConsumer {
-}
